fix(front): Unify to be in the same file hierarchy
Maybe confrict

Refs: #169
