+++ client/src/components/layout/layout.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import Header from "~/components/header";
-import Footer from "~/components/footer";
-import style from "~/styles/layout.module.scss";
+import Header from "~/components/layout/header";
+import Footer from "~/components/layout/footer";
+import style from "~/styles/components/layout/layout.module.scss";
 
 type Props = {
   children: React.ReactNode;
