+++ client/src/stories/layout/footer.stories.tsx
@@ -1,13 +1,11 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { ComponentMeta, ComponentStoryObj } from "@storybook/react";
-import Footer from "~/components/footer";
+import Footer from "~/components/layout/footer";
 
 type T = typeof Footer;
 type Story = ComponentStoryObj<T>;
 
 export default {
-  title: "Footer",
-
   component: Footer,
 } as ComponentMeta<T>;
