+++ client/src/stories/layout/header.stories.tsx
@@ -1,0 +1,12 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { ComponentMeta, ComponentStoryObj } from "@storybook/react";
+import Header from "~/components/layout/header";
+
+type T = typeof Header;
+type Story = ComponentStoryObj<T>;
+
+export default {
+  component: Header,
+} as ComponentMeta<T>;
+
+export const Default: Story = {};
