+++ client/src/stories/layout/layout.stories.tsx
@@ -1,0 +1,18 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { ComponentMeta, ComponentStoryObj } from "@storybook/react";
+import Layout from "~/components/layout/layout";
+import React from "react";
+
+type T = typeof Layout;
+type Story = ComponentStoryObj<T>;
+
+const data = {
+  node: <div>hogehoge</div>,
+};
+
+export default {
+  args: { children: data.node },
+  component: Layout,
+} as ComponentMeta<T>;
+
+export const Default: Story = {};
