+++ client/src/styles/components/button/anchor-button.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as var;
+@use "../../variables" as var;
 
 .enshita-button-icon-box {
   display: flex;
