+++ client/src/styles/components/layout/footer.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as var;
+@use "../../_variables" as var;
 
 .enshita-footer {
   box-sizing: border-box;
