+++ client/src/styles/components/layout/header.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as var;
+@use "../../variables" as var;
 
 .enshita-header {
   display: flex;
