+++ client/src/tests/components/button/anchor-text.test.tsx
@@ -1,26 +1,18 @@
 import { composeStories } from "@storybook/testing-react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import * as stories from "~/stories/button.stories";
+import * as stories from "~/stories/button/anchor-button.stories";
 
-const { Default, Disabled, Icon } = composeStories(stories);
+const { Default, Icon } = composeStories(stories);
 
-describe("(components) button", () => {
+describe("(components) anchor button", () => {
   const options = { name: "This is test" };
   test("[role=button]である", () => {
     const { getByRole } = render(<Default />);
     expect(getByRole("button", options)).toBeInTheDocument();
   });
-  test("[disabled=true]である", () => {
-    const { getByRole } = render(<Disabled />);
-    expect(getByRole("button", options)).toBeDisabled();
-  });
   test("(normal) test Snap Shot", () => {
     const { container } = render(<Default />);
-    expect(container).toMatchSnapshot();
-  });
-  test("(disable) test Snap Shot", () => {
-    const { container } = render(<Disabled />);
     expect(container).toMatchSnapshot();
   });
   test("(icon) test Snap Shot", () => {
