+++ client/src/tests/components/layout/footer.test.tsx
@@ -1,7 +1,7 @@
 import { composeStories } from "@storybook/testing-react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import * as stories from "~/stories/footer.stories";
+import * as stories from "~/stories/layout/footer.stories";
 
 const { Default } = composeStories(stories);
